requirements.txt
aisuite
tavily-python
jinja2
pydantic
openai
huggingface_hub
transformers
torch
python-dotenv